split promotion tests
